requirements.txt
# requirements.txt
# PDF OCR Text Extractor - Python Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
authlib==1.3.0
itsdangerous==2.1.2

# OCR & Image Processing
pytesseract==0.3.10
opencv-python==4.9.0.80
Pillow==10.2.0
pdf2image==1.17.0
numpy==1.26.3

# Email
aiosmtplib==3.0.1

# Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client (for OAuth)
httpx==0.26.0
requests==2.31.0

# Redis (optional for sessions)
redis==5.0.1
hiredis==2.3.2

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0